lint

+++ gnovm/stdlibs/time/time.go
@@ -13,5 +13,5 @@ func X_now(m *gno.Machine) (sec int64, nsec int32, mono int64) {
 	}
 
 	ctx := m.Context.(std.ExecContext)
-	return ctx.Timestamp, int32(ctx.TimestampNano), ctx.Timestamp*int64(time.Second) + int64(ctx.TimestampNano)
+	return ctx.Timestamp, int32(ctx.TimestampNano), ctx.Timestamp*int64(time.Second) + ctx.TimestampNano
 }
